+++ cli/cli/helpers/grafloki/kubernetes_grafloki.go
@@ -1,0 +1,8 @@
+package grafloki
+
+import "context"
+
+func StartGrafLokiInKubernetes(ctx context.Context) (string, error) {
+
+	return "", nil
+}
